backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Settings & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LangChain & LangGraph
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langgraph>=0.0.20

# Supabase
supabase>=2.3.0

# JWT
PyJWT>=2.8.0

# HTTP Client
httpx>=0.26.0

# Web Search
tavily-python>=0.3.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
